✨ Refactor: Simplify application lifespan management by moving logic to events module and updating related handlers
